requirements.txt
# Audio capture and processing
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Local speech-to-text
faster-whisper>=1.0.0

# LLM summarisation
anthropic>=0.40.0

# Notion integration
notion-client>=2.2.0

# Configuration
pyyaml>=6.0.1

# Utilities
python-slugify>=8.0.0
